+++ core/kernel/sub.mk
@@ -39,6 +39,7 @@ srcs-y += mutex.c
 srcs-$(CFG_LOCKDEP) += mutex_lockdep.c
 srcs-y += wait_queue.c
 srcs-y += notif.c
+srcs-$(_CFG_CORE_ASYNC_NOTIF_DEFAULT_IMPL) += notif_default.c
 srcs-y += thread.c
 
 ifeq ($(CFG_WITH_USER_TA),y)
